+++ src/main/java/model/participant/Name.java
@@ -13,7 +13,6 @@ public class Name {
     private void validate(String inputName) {
         validateBlank(inputName);
         validateContainBlankInName(inputName);
-        validateAllowedName(inputName);
     }
 
     private void validateBlank(String initialInput) {
@@ -25,12 +24,6 @@ public class Name {
     private void validateContainBlankInName(String inputName) {
         if (inputName.contains(" ")) {
             throw new IllegalArgumentException("중간에 공백이 포함된 값은 입력할 수 없습니다.");
-        }
-    }
-
-    private void validateAllowedName(String inputName){
-        if (inputName.contains("딜러")){
-            throw new IllegalArgumentException("'딜러'가 포함된 사용자명은 사용할 수 없습니다.");
         }
     }
