+++ src/test/java/model/participant/NameTest.java
@@ -22,11 +22,4 @@ class NameTest {
     void validate_ShouldThrowException_WhenInputStringContainsBlank(String value) {
         assertThrows(IllegalArgumentException.class, () -> new Name(value));
     }
-
-    @ParameterizedTest
-    @DisplayName("'딜러'가 포함된 문자열을 이용한 생성은 예외를 발생한다.")
-    @ValueSource(strings = {"딜러", "딜러도비", "딜러꼬붕", "진짜딜러"})
-    void validate_ShouldThrowException_WhenInputStringContainsDealerName(String value) {
-        assertThrows(IllegalArgumentException.class, () -> new Name(value));
-    }
 }
